+++ community/src/main/java/com/aisw/community/service/post/notice/CouncilApiLogicService.java
@@ -1,5 +1,7 @@
 package com.aisw.community.service.post.notice;
 
+import com.aisw.community.advice.exception.NotEqualAccountException;
+import com.aisw.community.advice.exception.PostNotFoundException;
 import com.aisw.community.advice.exception.UserNotFoundException;
 import com.aisw.community.model.entity.user.Account;
 import com.aisw.community.model.entity.post.notice.Council;
@@ -40,7 +42,8 @@ public class CouncilApiLogicService extends NoticePostService<CouncilApiRequest,
     @Override
     public Header<CouncilApiResponse> create(Header<CouncilApiRequest> request) {
         CouncilApiRequest councilApiRequest = request.getData();
-        Account account = accountRepository.findById(councilApiRequest.getAccountId()).orElseThrow(UserNotFoundException::new);
+        Account account = accountRepository.findById(councilApiRequest.getAccountId()).orElseThrow(
+                () -> new UserNotFoundException(councilApiRequest.getAccountId()));
         Council council = Council.builder()
                 .title(councilApiRequest.getTitle())
                 .writer(account.getName())
@@ -64,7 +67,7 @@ public class CouncilApiLogicService extends NoticePostService<CouncilApiRequest,
                 .map(council -> baseRepository.save((Council)council))
                 .map(this::response)
                 .map(Header::OK)
-                .orElseGet(() -> Header.ERROR("데이터 없음"));
+                .orElseThrow(() -> new PostNotFoundException(id));
     }
 
     @Override
@@ -72,28 +75,32 @@ public class CouncilApiLogicService extends NoticePostService<CouncilApiRequest,
     public Header<CouncilApiResponse> update(Header<CouncilApiRequest> request) {
         CouncilApiRequest councilApiRequest = request.getData();
 
-        return baseRepository.findById(councilApiRequest.getId())
-                .map(council -> {
-                    council
-                            .setTitle(councilApiRequest.getTitle())
-                            .setContent(councilApiRequest.getContent())
-                            .setStatus(councilApiRequest.getStatus());
-                    return council;
-                })
-                .map(council -> baseRepository.save(council))
-                .map(this::response)
-                .map(Header::OK)
-                .orElseGet(() -> Header.ERROR("데이터 없음"));
+        Council council = baseRepository.findById(councilApiRequest.getId()).orElseThrow(
+                () -> new PostNotFoundException(councilApiRequest.getId()));
+
+        if(council.getAccount().getId() != councilApiRequest.getAccountId()) {
+            throw new NotEqualAccountException(councilApiRequest.getAccountId());
+        }
+
+        council
+                .setTitle(councilApiRequest.getTitle())
+                .setContent(councilApiRequest.getContent())
+                .setStatus(councilApiRequest.getStatus());
+        baseRepository.save(council);
+
+        return Header.OK(response(council));
     }
 
     @Override
-    public Header delete(Long id) {
-        return baseRepository.findById(id)
-                .map(council -> {
-                    baseRepository.delete(council);
-                    return Header.OK();
-                })
-                .orElseGet(() -> Header.ERROR("데이터 없음"));
+    public Header delete(Long id, Long userId) {
+        Council council = baseRepository.findById(id).orElseThrow(() -> new PostNotFoundException(id));
+
+        if (council.getAccount().getId() != userId) {
+            throw new NotEqualAccountException(userId);
+        }
+
+        baseRepository.delete(council);
+        return Header.OK();
     }
 
     private CouncilApiResponse response(Council council) {
@@ -116,20 +123,7 @@ public class CouncilApiLogicService extends NoticePostService<CouncilApiRequest,
     }
 
     @Override
-    public void crawling(Long boardNo) throws IOException {
-    }
-
-    @Override
-    public Header<CouncilApiResponse> write(MultipartFile[] files) {
-        return null;
-    }
-
-    @Override
-    public ResponseEntity<Resource> download(Long id, String originFileName) {
-        return null;
-    }
-
-    @Override
+    @Cacheable(value = "councilSearch", key = "#pageable.pageNumber")
     public Header<NoticeResponseDTO> search(Pageable pageable) {
         Page<Council> councils = baseRepository.findAll(pageable);
         Page<Council> councilsByStatus = searchByStatus(pageable);
