+++ community/src/main/java/com/aisw/community/service/post/notice/UniversityApiLogicService.java
@@ -1,5 +1,7 @@
 package com.aisw.community.service.post.notice;
 
+import com.aisw.community.advice.exception.NotEqualAccountException;
+import com.aisw.community.advice.exception.PostNotFoundException;
 import com.aisw.community.advice.exception.UserNotFoundException;
 import com.aisw.community.model.entity.user.Account;
 import com.aisw.community.model.entity.post.attachment.Attachment;
@@ -41,6 +43,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
@@ -58,7 +61,8 @@ public class UniversityApiLogicService extends NoticePostService<UniversityApiRe
     @Override
     public Header<UniversityApiResponse> create(Header<UniversityApiRequest> request) {
         UniversityApiRequest universityApiRequest = request.getData();
-        Account account = accountRepository.findById(universityApiRequest.getAccountId()).orElseThrow(UserNotFoundException::new);
+        Account account = accountRepository.findById(universityApiRequest.getAccountId()).orElseThrow(
+                () -> new UserNotFoundException(universityApiRequest.getAccountId()));
         University university = University.builder()
                 .title(universityApiRequest.getTitle())
                 .writer(account.getName())
@@ -83,7 +87,7 @@ public class UniversityApiLogicService extends NoticePostService<UniversityApiRe
                 .map(university -> baseRepository.save((University) university))
                 .map(this::response)
                 .map(Header::OK)
-                .orElseGet(() -> Header.ERROR("데이터 없음"));
+                .orElseThrow(() -> new PostNotFoundException(id));
     }
 
     @Override
@@ -91,42 +95,46 @@ public class UniversityApiLogicService extends NoticePostService<UniversityApiRe
     public Header<UniversityApiResponse> update(Header<UniversityApiRequest> request) {
         UniversityApiRequest universityApiRequest = request.getData();
 
-        return baseRepository.findById(universityApiRequest.getId())
-                .map(university -> {
-                    university
-                            .setTitle(universityApiRequest.getTitle())
-                            .setContent(universityApiRequest.getContent())
-                            .setStatus(universityApiRequest.getStatus());
-                    university.setCampus(universityApiRequest.getCampus());
-                    return university;
-                })
-                .map(university -> baseRepository.save(university))
-                .map(this::response)
-                .map(Header::OK)
-                .orElseGet(() -> Header.ERROR("데이터 없음"));
+        University university = baseRepository.findById(universityApiRequest.getId()).orElseThrow(
+                () -> new PostNotFoundException(universityApiRequest.getId()));
+
+        if(university.getAccount().getId() != universityApiRequest.getAccountId()) {
+            throw new NotEqualAccountException(universityApiRequest.getAccountId());
+        }
+
+        university
+                .setTitle(universityApiRequest.getTitle())
+                .setContent(universityApiRequest.getContent())
+                .setStatus(universityApiRequest.getStatus());
+        university.setCampus(universityApiRequest.getCampus());
+        baseRepository.save(university);
+
+        return Header.OK(response(university));
     }
 
     @Override
-    public Header delete(Long id) {
-        return baseRepository.findById(id)
-                .map(university -> {
-                    baseRepository.delete(university);
-                    return Header.OK();
-                })
-                .orElseGet(() -> Header.ERROR("데이터 없음"));
+    public Header delete(Long id, Long userId) {
+        University university = baseRepository.findById(id).orElseThrow(() -> new PostNotFoundException(id));
+
+        if (university.getAccount().getId() != userId) {
+            throw new NotEqualAccountException(userId);
+        }
+
+        baseRepository.delete(university);
+        return Header.OK();
     }
 
-    @Override
-    public void crawling(Long boardNo) throws IOException{
-        Document doc = Jsoup.connect("https://www.gachon.ac.kr/community/opencampus/03.jsp?mode=view&boardType_seq=358&board_no=" + boardNo.toString()).get();
-        Elements elements = doc.select("table.view").select("tbody").select("tr");
-
-        String title = elements.get(0).select("td").text();
-        String writer = elements.get(1).select("td").get(0).text();
-        String createdAt = elements.get(2).select("td").text();
-        String campus = elements.get(3).select("td").text();
-        Elements files = elements.get(4).select("td a");
-        String contentHtml = elements.get(elements.size() - 1).toString();
+//    @Override
+//    public void crawling(Long boardNo) throws IOException{
+//        Document doc = Jsoup.connect("https://www.gachon.ac.kr/community/opencampus/03.jsp?mode=view&boardType_seq=358&board_no=" + boardNo.toString()).get();
+//        Elements elements = doc.select("table.view").select("tbody").select("tr");
+//
+//        String title = elements.get(0).select("td").text();
+//        String writer = elements.get(1).select("td").get(0).text();
+//        String createdAt = elements.get(2).select("td").text();
+//        String campus = elements.get(3).select("td").text();
+//        Elements files = elements.get(4).select("td a");
+//        String contentHtml = elements.get(elements.size() - 1).toString();
 
 //        System.out.println(title);
 //        System.out.println(writer);
@@ -137,87 +145,87 @@ public class UniversityApiLogicService extends NoticePostService<UniversityApiRe
 //            System.out.println("gachon.ac.kr" + file.attr("href"));
 //        }
 //        System.out.println(contentHtml);
+//
+//        Campus univ;
+//        if(campus.equals("공통"))
+//            univ = Campus.COMMON;
+//        else if(campus.equals("글로벌"))
+//            univ = Campus.GLOBAL;
+//        else
+//            univ = Campus.MEDICAL;
+//
+//        Account account = accountRepository.findById(1L).orElseThrow(() -> new UserNotFoundException(1L));
+//        University university = University.builder()
+//                .title(title)
+//                .writer(writer)
+//                .content(contentHtml)
+//                .status(BulletinStatus.GENERAL)
+//                .views(0L)
+//                .campus(univ)
+//                .firstCategory(FirstCategory.NOTICE)
+//                .secondCategory(SecondCategory.UNIVERSITY)
+//                .account(account)
+//                .build();
+//
+//        University newUniversity = baseRepository.save(university);
+//
+//        List<Attachment> attachmentList = new ArrayList<>();
+//        for(Element file : files){
+//            Attachment attachment = Attachment.builder()
+//                    .originFileName(file.text())
+//                    .fileName(file.text())
+//                    .filePath("gachon.ac.kr" + file.attr("href"))
+//                    .fileType(file.text())
+//                    .fileSize(1L)
+//                    .bulletin(newUniversity)
+//                    .build();
+//
+//            attachmentList.add(attachment);
+//            attachmentRepository.save(attachment);
+//        }
+//    }
 
-        Campus univ;
-        if(campus.equals("공통"))
-            univ = Campus.COMMON;
-        else if(campus.equals("글로벌"))
-            univ = Campus.GLOBAL;
-        else
-            univ = Campus.MEDICAL;
-
-        Account account = accountRepository.findById(1L).orElseThrow(UserNotFoundException::new);
-        University university = University.builder()
-                .title(title)
-                .writer(writer)
-                .content(contentHtml)
-                .status(BulletinStatus.GENERAL)
-                .views(0L)
-                .campus(univ)
-                .firstCategory(FirstCategory.NOTICE)
-                .secondCategory(SecondCategory.UNIVERSITY)
-                .account(account)
-                .build();
-
-        University newUniversity = baseRepository.save(university);
-
-        List<Attachment> attachmentList = new ArrayList<>();
-        for(Element file : files){
-            Attachment attachment = Attachment.builder()
-                    .originFileName(file.text())
-                    .fileName(file.text())
-                    .filePath("gachon.ac.kr" + file.attr("href"))
-                    .fileType(file.text())
-                    .fileSize(1L)
-                    .bulletin(newUniversity)
-                    .build();
-
-            attachmentList.add(attachment);
-            attachmentRepository.save(attachment);
-        }
-    }
-
-    @Override
-    public Header<UniversityApiResponse> write(MultipartFile[] files) throws IOException{
-        Account account = accountRepository.findById(1L).orElseThrow(UserNotFoundException::new);
-        University university = University.builder()
-                .title("Test01")
-                .writer("Test01")
-                .content("Content01")
-                .status(BulletinStatus.GENERAL)
-                .views(0L)
-                .campus(Campus.GLOBAL)
-                .firstCategory(FirstCategory.NOTICE)
-                .secondCategory(SecondCategory.UNIVERSITY)
-                .account(account)
-                .build();
-
-        University newUniversity = baseRepository.save(university);
-
-        List<Attachment> attachmentList = new ArrayList<>();
-        for(MultipartFile file : files) {
-            File destinationFile;
-            String fileName = file.getOriginalFilename();
-            String fileNameExtension = FilenameUtils.getExtension(fileName).toLowerCase();
-            String destinationFileName;
-
-            destinationFileName = RandomStringUtils.randomAlphanumeric(32) + "." + fileNameExtension;
-            destinationFile = new File("/Users/wonchang/desktop/files/" + destinationFileName);
-            destinationFile.getParentFile().mkdirs();
-            file.transferTo(destinationFile);
-
-            Attachment attachment = Attachment.builder()
-                    .originFileName(fileName)
-                    .fileName(destinationFileName)
-                    .filePath(destinationFile.getAbsolutePath())
-                    .fileType(file.getContentType())
-                    .fileSize(file.getSize())
-                    .bulletin(newUniversity)
-                    .build();
-
-            Attachment newAttachment = attachmentRepository.save(attachment);
-            attachmentList.add(newAttachment);
-        }
+//    @Override
+//    public Header<UniversityApiResponse> write(MultipartFile[] files) throws IOException{
+//        Account account = accountRepository.findById(1L).orElseThrow(UserNotFoundException::new);
+//        University university = University.builder()
+//                .title("Test01")
+//                .writer("Test01")
+//                .content("Content01")
+//                .status(BulletinStatus.GENERAL)
+//                .views(0L)
+//                .campus(Campus.GLOBAL)
+//                .firstCategory(FirstCategory.NOTICE)
+//                .secondCategory(SecondCategory.UNIVERSITY)
+//                .account(account)
+//                .build();
+//
+//        University newUniversity = baseRepository.save(university);
+//
+//        List<Attachment> attachmentList = new ArrayList<>();
+//        for(MultipartFile file : files) {
+//            File destinationFile;
+//            String fileName = file.getOriginalFilename();
+//            String fileNameExtension = FilenameUtils.getExtension(fileName).toLowerCase();
+//            String destinationFileName;
+//
+//            destinationFileName = RandomStringUtils.randomAlphanumeric(32) + "." + fileNameExtension;
+//            destinationFile = new File("/Users/wonchang/desktop/files/" + destinationFileName);
+//            destinationFile.getParentFile().mkdirs();
+//            file.transferTo(destinationFile);
+//
+//            Attachment attachment = Attachment.builder()
+//                    .originFileName(fileName)
+//                    .fileName(destinationFileName)
+//                    .filePath(destinationFile.getAbsolutePath())
+//                    .fileType(file.getContentType())
+//                    .fileSize(file.getSize())
+//                    .bulletin(newUniversity)
+//                    .build();
+//
+//            Attachment newAttachment = attachmentRepository.save(attachment);
+//            attachmentList.add(newAttachment);
+//        }
 //        try {
 //            fileName = new MD5Generator(file.getOriginalFilename()).toString();
 //        }catch (Exception e){
@@ -227,37 +235,37 @@ public class UniversityApiLogicService extends NoticePostService<UniversityApiRe
 //        destinationFile = new File("/Users/wonchang/desktop/files/" + fileName);
 //        destinationFile.getParentFile().mkdirs();
 //        file.transferTo(destinationFile);
+//
+//        return Header.OK(response(newUniversity));
+//    }
 
-        return Header.OK(response(newUniversity));
-    }
-
-    @Override
-    public ResponseEntity<Resource> download(Long id, String originFileName) throws IOException{
-        University university = universityRepository.findById(id).orElseThrow(UserNotFoundException::new);
-
-        List<Attachment> attachmentList = university.getAttachment();
-
-        for(Attachment attachment : attachmentList){
-            if(attachment.getOriginFileName().equals(originFileName)){
-                System.out.println("=============== Same ===============");
-                System.out.println("Original File Name : " + attachment.getOriginFileName());
-                System.out.println("Stored File Name : " + attachment.getFileName());
-                System.out.println("File Type : " + attachment.getFileType());
-                System.out.println("File Path : " + attachment.getFilePath());
-                System.out.println("File Size : " + attachment.getFileSize());
-
-                Path filePath = Paths.get(attachment.getFilePath()).toAbsolutePath().normalize();
-                Resource resource = new UrlResource(filePath.toUri());
-
-                return ResponseEntity.ok()
-                        .contentType(MediaType.APPLICATION_OCTET_STREAM)
-                        .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + attachment.getOriginFileName() + "\"")
-                        .body(resource);
-            }
-        }
-
-        return null;
-    }
+//    @Override
+//    public ResponseEntity<Resource> download(Long id, String originFileName) throws IOException{
+//        University university = universityRepository.findById(id).orElseThrow(UserNotFoundException::new);
+//
+//        Set<Attachment> attachmentList = university.getAttachment();
+//
+//        for(Attachment attachment : attachmentList){
+//            if(attachment.getOriginFileName().equals(originFileName)){
+//                System.out.println("=============== Same ===============");
+//                System.out.println("Original File Name : " + attachment.getOriginFileName());
+//                System.out.println("Stored File Name : " + attachment.getFileName());
+//                System.out.println("File Type : " + attachment.getFileType());
+//                System.out.println("File Path : " + attachment.getFilePath());
+//                System.out.println("File Size : " + attachment.getFileSize());
+//
+//                Path filePath = Paths.get(attachment.getFilePath()).toAbsolutePath().normalize();
+//                Resource resource = new UrlResource(filePath.toUri());
+//
+//                return ResponseEntity.ok()
+//                        .contentType(MediaType.APPLICATION_OCTET_STREAM)
+//                        .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + attachment.getOriginFileName() + "\"")
+//                        .body(resource);
+//            }
+//        }
+//
+//        return null;
+//    }
 
     private UniversityApiResponse response(University university) {
         UniversityApiResponse universityApiResponse = UniversityApiResponse.builder()
